fix recursive deps
